Move error rescue notification function in notification center
