+++ lib/kite-editor.js
@@ -149,7 +149,7 @@ class KiteEditor {
             }
           } else {
             if (versionChanged) {
-              this.notifyCodeCleanUpVersionChange(data.version);
+              Kite.notifications.notifyErrorRescueVersionChange(data.version);
             }
           }
         }
@@ -159,23 +159,6 @@ class KiteEditor {
 
   mustOpenSidebar() {
     return atom.config.get('kite.actionWhenErrorRescueFixesCode') === AUTOCORRECT_SHOW_SIDEBAR;
-  }
-
-  notifyCodeCleanUpVersionChange(version) {
-    DataLoader.getAutocorrectModelInfo(version).then(data => {
-      const [example] = data.examples;
-      let notification = atom.notifications.addInfo(`#### Kite Error Rescue has just been updated\n${example.synopsis}`, {
-        dismissable: true,
-        detail: `${example.old.map(d => d.text).join('\n')}\n${example.new.map(d => d.text).join('\n')}`,
-        buttons: [{
-          text: 'Learn more',
-          onDidClick() {
-            notification.dismiss();
-            Kite.toggleAutocorrectSidebar(true);
-          },
-        }],
-      });
-    });
   }
 
   expandRange(range) {
