+++ lib/notifications-center.js
@@ -4,7 +4,7 @@ const os = require('os');
 const {CompositeDisposable, Emitter} = require('atom');
 const {Logger, StateController} = require('kite-installer');
 const {STATES} = StateController;
-const metrics = require('./metrics.js');
+const DataLoader = require('./data-loader');
 
 class NotificationsCenter {
   get NOTIFIERS() {
@@ -93,6 +93,26 @@ class NotificationsCenter {
   notify(state) {
     this[this.NOTIFIERS[state]] && this[this.NOTIFIERS[state]](state);
   }
+
+  notifyErrorRescueVersionChange(version) {
+    DataLoader.getAutocorrectModelInfo(version).then(data => {
+      const [example] = data.examples;
+      let notification = atom.notifications.addInfo(`#### Kite Error Rescue has just been updated\n${example.synopsis}`, {
+        dismissable: true,
+        detail: `${example.old.map(d => d.text).join('\n')}\n${example.new.map(d => d.text).join('\n')}`,
+        buttons: [{
+          text: 'Learn more',
+          onDidClick() {
+            const Kite = require('./kite');
+
+            notification.dismiss();
+            Kite.toggleAutocorrectSidebar(true);
+          },
+        }],
+      });
+    });
+  }
+
 
   warnNotSupported(state) {
     let description = 'Sorry, the Kite engine only supports macOS and Windows at the moment.';
